Refactor useBorrowTokensList and useCollateralsList hooks
