+++ src/pages/tokenLending/BorrowTokenPage/InstantBorrowContent/hooks/constants.ts
@@ -1,3 +1,1 @@
-export const DEFAULT_COLLATERAL_MINT = 'BANXbTpN8U2cU41FjPxe2Ti37PiT5cCxLUKDQZuJeMMR'
-
 export const MAX_TOKEN_TO_GET_TRESHOLD = 100
